third_party/WebKit/LayoutTests/fast/css/usecounter-before.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style id="style"></style>
<script>
'use strict';

test(() => {
    var kHasBeforeOrAfterPseudoElement = 2204; // From UseCounter.h
    var style = document.getElementById('style');
	style.textContent = "div::first-line { color: red; }";
    assert_false(internals.isUseCounted(document, kHasBeforeOrAfterPseudoElement));
	style.textContent = "div::before { color: red; }";
    assert_true(internals.isUseCounted(document, kHasBeforeOrAfterPseudoElement));
}, 'Use of ::before in CSS is use counted.');
</script>
